Refactory the page to fixes the memory leak error
